Move DSL macros to a separate package and split ChimneyBlackbox macros into TransformerBlackboxMacros and PatcherBlackboxMacros
